+++ src/components/Slider/Slider.scss
@@ -1,1 +1,98 @@
+.slider {
+  position: relative;
+  margin-top: 200px;
+  margin-bottom: 200px;
 
+  &:hover {
+    .leftArrow {
+      opacity: 1;
+      margin-left: 0;
+    }
+    .rightArrow {
+      opacity: 1;
+      margin-right: 0;
+    }
+  }
+
+  .leftArrow {
+    position: absolute;
+    width: 80px;
+    left: 0;
+    top: 40%;
+    margin-left: -80px;
+    opacity: 0;
+    transition: all 0.5s ease-in-out;
+    z-index: 1;
+  }
+  .rightArrow {
+    position: absolute;
+    width: 80px;
+    right: 0;
+    top: 40%;
+    margin-right: -80px;
+    opacity: 0;
+    transition: all 0.5s ease-in-out;
+    z-index: 1;
+  }
+
+  .carouselWrapper {
+    position: relative;
+    width: 1200px;
+    height: 550px;
+    margin: 0 auto;
+    overflow-x: scroll;
+
+    &::-webkit-scrollbar {
+      height: 3px;
+    }
+    &::-webkit-scrollbar-thumb {
+      height: auto;
+      background-color: #333333;
+    }
+    &::-webkit-scrollbar-track {
+      background-color: #cccbc1;
+    }
+
+    .carousel {
+      display: flex;
+      position: absolute;
+      transition: 0.5s;
+
+      .imgBox {
+        display: flex;
+        flex-direction: column;
+        align-items: center;
+        height: 100%;
+
+        .image {
+          height: 350px;
+          margin-top: 50px;
+        }
+        .image:hover ~ .text .recommendedItemName {
+          text-decoration: underline;
+          text-decoration-thickness: 1px;
+          cursor: pointer;
+        }
+      }
+      .recommendTitleBox {
+        width: 480px;
+        height: 400px;
+        margin-left: 20px;
+
+        p {
+          font-size: 30px;
+          color: #333333;
+          font-weight: 400;
+        }
+      }
+      .text {
+        margin-top: 15px;
+
+        p {
+          margin: 10px 0 10px 0;
+          font-weight: 300;
+        }
+      }
+    }
+  }
+}
